+++ modules/chowdsp_dsp/EQ/chowdsp_EQBand.h
@@ -60,11 +60,11 @@ public:
     void process (const ProcessContext& context);
 
 private:
-    template <typename FilterType, typename T = FloatType, int N = FilterType::Order>
+    template <typename FilterType, typename T = FloatType, size_t N = FilterType::Order>
     std::enable_if_t<std::is_base_of_v<IIRFilter<N, T>, FilterType>, void>
         processFilterChannel (FilterType& filter, juce::dsp::AudioBlock<FloatType>& block);
 
-    template <typename FilterType, typename T = FloatType, int N = FilterType::Order, StateVariableFilterType type = FilterType::Type>
+    template <typename FilterType, typename T = FloatType, size_t N = FilterType::Order, StateVariableFilterType type = FilterType::Type>
     std::enable_if_t<std::is_base_of_v<NthOrderFilter<T, N, type>, FilterType>, void>
         processFilterChannel (FilterType& filter, juce::dsp::AudioBlock<FloatType>& block);
